refactor: 클린 아키텍처 반영 #18
