+++ src/main/java/org/cheonyakplanet/be/application/dto/community/PostUpdateRequestDTO.java
@@ -1,6 +1,6 @@
 package org.cheonyakplanet.be.application.dto.community;
 
-import org.cheonyakplanet.be.domain.entity.comunity.PostCategory;
+import org.cheonyakplanet.be.domain.entity.community.PostCategory;
 
 import io.swagger.v3.oas.annotations.media.Schema;
