+++ src/main/java/org/cheonyakplanet/be/application/service/CommunityService.java
@@ -8,20 +8,20 @@ import org.cheonyakplanet.be.application.dto.community.PostCreateDTO;
 import org.cheonyakplanet.be.application.dto.community.PostDTO;
 import org.cheonyakplanet.be.application.dto.community.PostDetailDTO;
 import org.cheonyakplanet.be.application.dto.community.PostUpdateRequestDTO;
-import org.cheonyakplanet.be.domain.entity.comunity.Comment;
-import org.cheonyakplanet.be.domain.entity.comunity.Post;
-import org.cheonyakplanet.be.domain.entity.comunity.PostCategory;
-import org.cheonyakplanet.be.domain.entity.comunity.PostReaction;
-import org.cheonyakplanet.be.domain.entity.comunity.ReactionType;
-import org.cheonyakplanet.be.domain.entity.comunity.Reply;
+import org.cheonyakplanet.be.domain.entity.community.Comment;
+import org.cheonyakplanet.be.domain.entity.community.Post;
+import org.cheonyakplanet.be.domain.entity.community.PostCategory;
+import org.cheonyakplanet.be.domain.entity.community.PostReaction;
+import org.cheonyakplanet.be.domain.entity.community.ReactionType;
+import org.cheonyakplanet.be.domain.entity.community.Reply;
 import org.cheonyakplanet.be.domain.entity.user.User;
 import org.cheonyakplanet.be.domain.repository.CommentRepository;
 import org.cheonyakplanet.be.domain.repository.PostReactionRepository;
 import org.cheonyakplanet.be.domain.repository.PostRepository;
 import org.cheonyakplanet.be.domain.repository.ReplyRepository;
 import org.cheonyakplanet.be.infrastructure.security.UserDetailsImpl;
-import org.cheonyakplanet.be.presentation.exception.CustomException;
-import org.cheonyakplanet.be.presentation.exception.ErrorCode;
+import org.cheonyakplanet.be.domain.exception.CustomException;
+import org.cheonyakplanet.be.domain.exception.ErrorCode;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
