+++ src/main/java/org/cheonyakplanet/be/application/service/InfoService.java
@@ -26,9 +26,9 @@ import org.cheonyakplanet.be.domain.entity.subscription.SggCode;
 import org.cheonyakplanet.be.domain.repository.RealEstatePriceRepository;
 import org.cheonyakplanet.be.domain.repository.RealEstatePriceSummaryRepository;
 import org.cheonyakplanet.be.domain.repository.SggCodeRepository;
-import org.cheonyakplanet.be.domain.util.Utils;
-import org.cheonyakplanet.be.presentation.exception.CustomException;
-import org.cheonyakplanet.be.presentation.exception.ErrorCode;
+import org.cheonyakplanet.be.infrastructure.util.Utils;
+import org.cheonyakplanet.be.domain.exception.CustomException;
+import org.cheonyakplanet.be.domain.exception.ErrorCode;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.task.TaskExecutor;
