+++ src/main/java/org/cheonyakplanet/be/application/service/SubscriptionQueryService.java
@@ -8,17 +8,17 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import org.cheonyakplanet.be.application.dto.subscriprtion.SubscriptionDTO;
-import org.cheonyakplanet.be.application.dto.subscriprtion.SubscriptionDetailDTO;
-import org.cheonyakplanet.be.application.dto.subscriprtion.SubscriptionInfoSimpleDTO;
+import org.cheonyakplanet.be.application.dto.subscription.SubscriptionDTO;
+import org.cheonyakplanet.be.application.dto.subscription.SubscriptionDetailDTO;
+import org.cheonyakplanet.be.application.dto.subscription.SubscriptionInfoSimpleDTO;
 import org.cheonyakplanet.be.domain.entity.subscription.SubscriptionInfo;
 import org.cheonyakplanet.be.domain.entity.subscription.SubscriptionLocationInfo;
 import org.cheonyakplanet.be.domain.entity.user.User;
+import org.cheonyakplanet.be.domain.exception.CustomException;
+import org.cheonyakplanet.be.domain.exception.ErrorCode;
 import org.cheonyakplanet.be.domain.repository.SubscriptionInfoRepository;
 import org.cheonyakplanet.be.domain.repository.SubscriptionLocationInfoRepository;
 import org.cheonyakplanet.be.infrastructure.security.UserDetailsImpl;
-import org.cheonyakplanet.be.presentation.exception.CustomException;
-import org.cheonyakplanet.be.presentation.exception.ErrorCode;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
