--- src/main/java/org/cheonyakplanet/be/domain/entity/comunity/Comment.java
@@ -1,53 +1,0 @@
-package org.cheonyakplanet.be.domain.entity.comunity;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.cheonyakplanet.be.domain.Stamped;
-
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(catalog = "planet", name = "Comment")
-public class Comment extends Stamped {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column(columnDefinition = "TEXT")
-	private String content;
-
-	private String username;
-
-	@ManyToOne
-	@JsonBackReference
-	@JoinColumn(name = "post_id")
-	private Post post;
-
-	@Builder.Default
-	@JsonManagedReference
-	@OneToMany(mappedBy = "comment", cascade = CascadeType.ALL, orphanRemoval = true)
-	private List<Reply> replies = new ArrayList<>();
-}
