--- src/main/java/org/cheonyakplanet/be/domain/entity/comunity/PostCategory.java
@@ -1,9 +1,0 @@
-package org.cheonyakplanet.be.domain.entity.comunity;
-
-public enum PostCategory {
-	REVIEW,            // 후기
-	SUBSCRIPTION_INFO, // 청약정보
-	CHITCHAT,          // 잡담
-	INFO_SHARE,        // 정보공유
-	QUESTION           // 질문
-}
