--- src/main/java/org/cheonyakplanet/be/domain/entity/comunity/ReactionType.java
@@ -1,6 +1,0 @@
-package org.cheonyakplanet.be.domain.entity.comunity;
-
-public enum ReactionType {
-	LIKE,
-	DISLIKE
-}
