+++ src/main/java/org/cheonyakplanet/be/domain/repository/PostReactionRepository.java
@@ -2,8 +2,8 @@ package org.cheonyakplanet.be.domain.repository;
 
 import java.util.Optional;
 
-import org.cheonyakplanet.be.domain.entity.comunity.Post;
-import org.cheonyakplanet.be.domain.entity.comunity.PostReaction;
+import org.cheonyakplanet.be.domain.entity.community.Post;
+import org.cheonyakplanet.be.domain.entity.community.PostReaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostReactionRepository extends JpaRepository<PostReaction, Long> {
