+++ src/main/java/org/cheonyakplanet/be/infrastructure/jwt/JwtAuthorizationFilter.java
@@ -7,9 +7,9 @@ import org.cheonyakplanet.be.application.dto.ApiResponse;
 import org.cheonyakplanet.be.domain.entity.user.UserToken;
 import org.cheonyakplanet.be.domain.repository.UserTokenRepository;
 import org.cheonyakplanet.be.infrastructure.security.UserDetailsServiceImpl;
-import org.cheonyakplanet.be.presentation.exception.CustomException;
-import org.cheonyakplanet.be.presentation.exception.ErrorCode;
-import org.cheonyakplanet.be.presentation.exception.ErrorData;
+import org.cheonyakplanet.be.domain.exception.CustomException;
+import org.cheonyakplanet.be.domain.exception.ErrorCode;
+import org.cheonyakplanet.be.domain.exception.ErrorData;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
