+++ src/main/java/org/cheonyakplanet/be/presentation/controller/DataController.java
@@ -3,7 +3,7 @@ package org.cheonyakplanet.be.presentation.controller;
 import java.util.List;
 
 import org.cheonyakplanet.be.application.dto.ApiResponse;
-import org.cheonyakplanet.be.application.dto.subscriprtion.CoordinateResponseDTO;
+import org.cheonyakplanet.be.application.dto.subscription.CoordinateResponseDTO;
 import org.cheonyakplanet.be.application.service.FinanceService;
 import org.cheonyakplanet.be.application.service.InfoService;
 import org.cheonyakplanet.be.application.service.SubscriptionService;
