+++ src/main/java/org/cheonyakplanet/be/presentation/controller/InfoController.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.cheonyakplanet.be.application.dto.ApiResponse;
 import org.cheonyakplanet.be.application.dto.infra.InfraResponseDTO;
 import org.cheonyakplanet.be.application.dto.infra.PublicFacilityDTO;
-import org.cheonyakplanet.be.application.dto.subscriprtion.SubscriptionInfoSimpleDTO;
-import org.cheonyakplanet.be.application.dto.subscriprtion.SubscriptionLikeDTO;
+import org.cheonyakplanet.be.application.dto.subscription.SubscriptionInfoSimpleDTO;
+import org.cheonyakplanet.be.application.dto.subscription.SubscriptionLikeDTO;
 import org.cheonyakplanet.be.application.service.InfoService;
 import org.cheonyakplanet.be.application.service.InfrastructureService;
 import org.cheonyakplanet.be.application.service.SubscriptionLikeService;
