+++ src/main/java/org/cheonyakplanet/be/presentation/exception/CustomAuthenticationEntryPoint.java
@@ -1,43 +1,45 @@
 package org.cheonyakplanet.be.presentation.exception;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import java.io.IOException;
+
 import org.cheonyakplanet.be.application.dto.ApiResponse;
+import org.cheonyakplanet.be.domain.exception.ErrorData;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
-import java.io.IOException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @RequiredArgsConstructor
 @Component
 public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
-    private final ObjectMapper objectMapper;
+	private final ObjectMapper objectMapper;
 
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response,
-                         AuthenticationException authException) throws IOException {
+	@Override
+	public void commence(HttpServletRequest request, HttpServletResponse response,
+		AuthenticationException authException) throws IOException {
 
-        response.setStatus(HttpStatus.OK.value());
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setCharacterEncoding("UTF-8");
+		response.setStatus(HttpStatus.OK.value());
+		response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+		response.setCharacterEncoding("UTF-8");
 
-        ErrorData errorData = new ErrorData(
-                "AUTH001",
-                "인증이 필요한 서비스입니다",
-                authException.getMessage());
+		ErrorData errorData = new ErrorData(
+			"AUTH001",
+			"인증이 필요한 서비스입니다",
+			authException.getMessage());
 
-        ApiResponse<ErrorData> apiResponse = new ApiResponse<>("fail", errorData);
+		ApiResponse<ErrorData> apiResponse = new ApiResponse<>("fail", errorData);
 
-        String json = objectMapper.writeValueAsString(apiResponse);
-        response.getWriter().write(json);
-    }
+		String json = objectMapper.writeValueAsString(apiResponse);
+		response.getWriter().write(json);
+	}
 }
