--- src/main/java/org/cheonyakplanet/be/presentation/exception/ErrorData.java
@@ -1,15 +1,0 @@
-package org.cheonyakplanet.be.presentation.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter @Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ErrorData {
-    private String code;
-    private String message;
-    private String details;
-}
