+++ src/main/java/org/cheonyakplanet/be/presentation/exception/JwtExceptionFilter.java
@@ -6,6 +6,9 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.cheonyakplanet.be.application.dto.ApiResponse;
+import org.cheonyakplanet.be.domain.exception.CustomException;
+import org.cheonyakplanet.be.domain.exception.ErrorCode;
+import org.cheonyakplanet.be.domain.exception.ErrorData;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
